Simplify HermiteSimpsonSolver.cs by extracting constraint configuration
Extracted constraint configuration logic from HermiteSimpsonSolver into a new
ConstraintConfigurator helper class. This simplification improves code
organization and maintainability by:

- Creating a focused helper class for all constraint-related operations
- Reducing the HermiteSimpsonSolver class from 906 to 729 lines (-177 lines)
- Separating concerns between solving logic and constraint configuration
- Making constraint setup more modular and testable

All existing tests pass without modification.

Resolves #10
